ActivatableService fungere nu på samme måde som services du ville finde i laby's waypoints
