+++ core/src/main/java/dk/fvtrademarket/fvplus/core/activatable/DefaultGuardVault.java
@@ -1,19 +1,22 @@
-package dk.fvtrademarket.fvplus.api.activatable.misc;
+package dk.fvtrademarket.fvplus.core.activatable;
 
+import dk.fvtrademarket.fvplus.api.activatable.guardvault.GuardVault;
 import dk.fvtrademarket.fvplus.api.enums.FreakyVilleServer;
 import dk.fvtrademarket.fvplus.api.enums.PrisonSector;
 import net.labymod.api.client.component.Component;
 import net.labymod.api.client.component.TextComponent;
 import net.labymod.api.client.component.format.NamedTextColor;
 
-public class DefaultGangArea implements GangArea {
+public class DefaultGuardVault implements GuardVault {
   private final PrisonSector prisonSector;
+  private final PrisonSector[] visiblePrisonSectors;
 
   private final int expectedActivationTime, cooldownUponFailure, personalCooldown, sharedCooldown;
 
-  public DefaultGangArea(PrisonSector prisonSector,
+  public DefaultGuardVault(PrisonSector prisonSector, PrisonSector[] visiblePrisonSectors,
       int expectedActivationTime, int cooldownUponFailure, int personalCooldown, int sharedCooldown) {
     this.prisonSector = prisonSector;
+    this.visiblePrisonSectors = visiblePrisonSectors;
     this.expectedActivationTime = expectedActivationTime;
     this.cooldownUponFailure = cooldownUponFailure;
     this.personalCooldown = personalCooldown;
@@ -23,6 +26,11 @@ public class DefaultGangArea implements GangArea {
   @Override
   public PrisonSector getPrisonSector() {
     return this.prisonSector;
+  }
+
+  @Override
+  public PrisonSector[] getVisiblePrisonSectors() {
+    return this.visiblePrisonSectors;
   }
 
   @Override
@@ -52,10 +60,16 @@ public class DefaultGangArea implements GangArea {
 
   @Override
   public Component toComponent() {
+    String key = "fvplus.activatable.";
+    if (this.prisonSector == PrisonSector.A_PLUS || this.prisonSector == PrisonSector.B_PLUS) {
+      key += "bank";
+    } else {
+      key += "guardVault";
+    }
     return TextComponent.builder()
             .append(getPrisonSector().toComponent())
             .append(Component.space())
-            .append(Component.translatable("fvplus.activatable.gangArea", NamedTextColor.GRAY))
+            .append(Component.translatable(key, NamedTextColor.GRAY))
             .build();
   }
 }
